Fix messaging for cron log line
The current log line that gets emitted every cron evaluation is

```
running a task every minute
```

This is misleading as it's probably... not... running every minute. You'd probably run out of rate limit if you tried that.

This PR changes the log line to just output the cron line, for lack of another idea on what to output easily.

Please note that I made this PR in the github `/edit/` page, and have not executed this at all. Will try to patch our instance to test this change later today.