Don't get repo twice when handling default branch protection
When dealing with the default branch in branches plugin, the repo was read even though it had been already fetched.

This reduces the number of calls in a synch, it could be further optimized by only getting it if the default branch is reference. But not defining the default repo is very likely an edge case.