fix(repository): ignore field `security` (#388)
`security` is ignored from the comparison as it is not part of the
GitHub response. Using `security` in the config will no longer do
an update to the repo settings with updateRepo when there are no
changes to the repository settings.
updateSecurity is always called since the data it depends on is
not part of the repository response payload.